# Backend/api/openapi/teams-notification-api.yaml
openapi: 3.0.3
info:
  title: Teams Notification API (Go implementation)
  version: 1.0.0
  description: |
    This specification documents the REST endpoints implemented in the Teams Notification Go service.
    Includes all available routes for internal and external APIs.

    ## API Structure
    - **External APIs** (`/api/v1`): Public APIs for external clients
    - **Internal APIs** (`/internal/v1`): Internal management APIs
    - **System APIs**: Health, metrics, and configuration endpoints

    ## Recent Updates (v1.0.0)
    - **Unified Configuration**: Environment-based configuration system
    - **Error Handling**: Standardized error responses with proper HTTP status codes
    - **Token Management**: Redis-backed token caching for improved performance
    - **Monitoring**: Enhanced health checks and system metrics
    - **Logging**: Structured logging with configurable levels and formats
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.teamsnotify.com
    description: Production server
paths:
...
  /api/v1/destinations/{notifyKey}:
    get:
      summary: Get project destinations (External API)
      tags: [External API]
      description: Get destinations for a project using notify key
      parameters:
        - name: notifyKey
          in: path
          required: true
          schema:
            type: string
          description: Project notify key
      responses:
        "200":
          description: Project destinations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDestinationsResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"
...
  /internal/v1/destinations:
    get:
      summary: List destinations
      tags: [Internal API - Destinations]
      description: Get a list of destinations with pagination
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/SearchParam"
      responses:
        "200":
          description: List of destinations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationListResponse"

    post:
      summary: Create destination
      tags: [Internal API - Destinations]
      description: Create a new destination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationCreateRequest"
      responses:
        "201":
          description: Destination created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
  /internal/v1/destinations/{id}:
    get:
      summary: Get destination
      tags: [Internal API - Destinations]
      description: Get destination by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Destination details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"

    put:
      summary: Update destination
      tags: [Internal API - Destinations]
      description: Update destination by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationUpdateRequest"
      responses:
        "200":
          description: Destination updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    delete:
      summary: Delete destination
      tags: [Internal API - Destinations]
      description: Delete destination by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Destination deleted successfully
        "404":
          $ref: "#/components/responses/NotFoundError"
  /internal/v1/destinations/{id}/targets:
    patch:
      summary: Update destination targets
      tags: [Internal API - Destinations]
      description: Update destination targets by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationTargetsUpdateRequest"
      responses:
        "200":
          description: Destination targets updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /internal/v1/destinations/{id}/validate:
    post:
      summary: Validate destination targets
      tags: [Internal API - Destinations]
      description: Validate destination targets by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Destination targets validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationValidationResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /internal/v1/destinations/project/{projectId}:
    get:
      summary: Get destinations by project
      tags: [Internal API - Destinations]
      description: Get destinations by project ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of destinations for project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationListResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /internal/v1/destinations/bot/{botId}:
    get:
      summary: Get destinations by bot
      tags: [Internal API - Destinations]
      description: Get destinations by bot ID
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of destinations for bot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationListResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /internal/v1/destinations/search:
    get:
      summary: Search destinations
      tags: [Internal API - Destinations]
      description: Search destinations with filters
      parameters:
        - $ref: "#/components/parameters/SearchParam"
        - name: type
          in: query
          schema:
            type: string
            enum: [personal, group, channel]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationListResponse"

  # Bots endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    LimitParam:
      name: limit
      in: query
      description: Number of items to return
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    OffsetParam:
      name: offset
      in: query
      description: Number of items to skip
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    SearchParam:
      name: search
      in: query
      description: Search term
      required: false
      schema:
        type: string
...
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFoundError:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Detailed error information
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
...
    PaginationResponse:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        pages:
          type: integer

    # Bot Framework schemas
...
    ProjectDestinationsResponse:
      type: object
      properties:
        notifyKey:
          type: string
        destinations:
          type: array
          items:
            $ref: "#/components/schemas/DestinationResponse"
    DestinationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [personal, group, channel]
        conversationId:
          type: string
        displayName:
          type: string
        tenantId:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time

    # Configuration schemas
...
    DestinationCreateRequest:
      type: object
      required:
        - projectId
        - botId
        - type
        - targets
      properties:
        projectId:
          type: string
          format: uuid
          description: Project ID
        botId:
          type: string
          format: uuid
          description: Bot ID
        type:
          type: string
          enum: [personal, group, channel]
          description: Destination type
        targets:
          type: array
          items:
            $ref: "#/components/schemas/DestinationTarget"
          description: List of targets
        status:
          type: string
          enum: [active, inactive]
          default: active
          description: Destination status
    DestinationUpdateRequest:
      type: object
      properties:
        type:
          type: string
          enum: [personal, group, channel]
        targets:
          type: array
          items:
            $ref: "#/components/schemas/DestinationTarget"
        status:
          type: string
          enum: [active, inactive]
    DestinationTarget:
      type: object
      required:
        - type
        - conversation_id
        - display_name
        - tenant_id
      properties:
        type:
          type: string
          enum: [personal, group, channel]
        conversation_id:
          type: string
          description: Teams conversation ID
        display_name:
          type: string
          description: Display name
        tenant_id:
          type: string
          format: uuid
          description: Teams tenant ID
    DestinationTargetsUpdateRequest:
      type: object
      required:
        - targets
      properties:
        targets:
          type: array
          items:
            $ref: "#/components/schemas/DestinationTarget"
    DestinationValidationResponse:
      type: object
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
    DestinationListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/DestinationResponse"
        pagination:
          $ref: "#/components/schemas/PaginationResponse"

    # Bot schemas
...